+++ js/plugins/tabs-material.js
@@ -9,6 +9,7 @@ var MaterialTabsDefaultConfig = {
 
     onBeforeTabOpen: Metro.noop_true,
     onTabOpen: Metro.noop,
+    onTabsScroll: Metro.noop,
     onTabsCreate: Metro.noop
 };
 
@@ -26,6 +27,8 @@ var MaterialTabs = {
         this.elem  = elem;
         this.element = $(elem);
         this.marker = null;
+        this.scroll = 0;
+        this.scrollDir = "left";
 
         this._setOptionsFromDOM();
         this._create();
@@ -34,7 +37,7 @@ var MaterialTabs = {
     },
 
     _setOptionsFromDOM: function(){
-        var that = this, element = this.element, o = this.options;
+        var element = this.element, o = this.options;
 
         $.each(element.data(), function(key, value){
             if (key in o) {
@@ -48,7 +51,7 @@ var MaterialTabs = {
     },
 
     _create: function(){
-        var that = this, element = this.element, o = this.options;
+        var element = this.element, o = this.options;
 
         this._createStructure();
         this._createEvents();
@@ -71,7 +74,7 @@ var MaterialTabs = {
     },
 
     _createStructure: function(){
-        var that = this, element = this.element, o = this.options;
+        var element = this.element, o = this.options;
         var tabs = element.find("li"), active_tab = element.find("li.active");
 
         element.addClass("tabs-material").addClass(o.clsComponent);
@@ -114,21 +117,25 @@ var MaterialTabs = {
             }
         });
 
-        var addMouseWheel = function (){
-            $(element).on(Metro.events.mousewheel, function(e){
-                var scroll_value = e.deltaX * METRO_SCROLL_MULTIPLE;
-                element.scrollLeft(element.scrollLeft() - scroll_value);
-                return false;
-            });
-        };
+        element.on(Metro.events.scroll, function(){
+            var oldScroll = this.scroll;
 
-        if (!$('html').hasClass("metro-touch-device")) {
-            addMouseWheel();
-        }
+            this.scrollDir = this.scroll < element[0].scrollLeft ? "left" : "right";
+            this.scroll = element[0].scrollLeft;
+
+            Utils.exec(o.onTabsScroll, [element[0].scrollLeft, oldScroll, this.scrollDir], element[0]);
+
+            element.fire("tabsscroll", {
+                scrollLeft: element[0].scrollLeft,
+                oldScroll: oldScroll,
+                scrollDir: that.scrollDir
+            });
+
+        });
     },
 
     openTab: function(tab, tab_next){
-        var that = this, element = this.element, o = this.options;
+        var element = this.element, o = this.options;
         var tabs = element.find("li"), element_scroll = element.scrollLeft();
         var magic = 32, shift, width = element.width(), tab_width, target, tab_left;